v0.0.6-alpha

+++ CTS256A-AL2/CPU.h
@@ -46,7 +46,7 @@ public:
 	// set PC
 	void setPC( uint pc )
 	{
-		pc_ = pc;
+		pc_ = ushort(pc);
 	}
 
 	// get PC
